+++ src/GameLibrary/Exception/Utilities/ArrayListOutOfBoundsException.h
@@ -10,16 +10,16 @@ namespace GameLibrary
 	{
 	public:
 		/*! the out of bounds index*/
-		unsigned int index;
+		size_t index;
 		/*! the size of the ArrayList*/
-		unsigned int size;
+		size_t size;
 		
 		/*! copy constructor*/
 		ArrayListOutOfBoundsException(const ArrayListOutOfBoundsException&);
 		/*! Constructs an exception with an ArrayList index and size
 			\param index the index in the ArrayList
 			\param size the size of the ArrayList*/
-		ArrayListOutOfBoundsException(unsigned int index, unsigned int size);
+		ArrayListOutOfBoundsException(size_t index, size_t size);
 		/*! virtual destructor*/
 		virtual ~ArrayListOutOfBoundsException();
 		/*! assignment operator*/
