+++ src/GameLibrary/Exception/Utilities/BitSetOutOfBoundsException.h
@@ -10,13 +10,13 @@ namespace GameLibrary
 	{
 	public:
 		/*! the out of bounds index*/
-		unsigned int index;
+		size_t index;
 		
 		/*! copy constructor*/
 		BitSetOutOfBoundsException(const BitSetOutOfBoundsException&);
 		/*! Constructs an exception with the BitSet out of bounds index.
 			\param index the index in the BitSet*/
-		BitSetOutOfBoundsException(unsigned int index);
+		BitSetOutOfBoundsException(size_t index);
 		/*! virtual destructor*/
 		virtual ~BitSetOutOfBoundsException();
 		/*! assignment operator*/
